requirements.txt
streamlit
python-dotenv
requests
together  # If you're using the official Together Python SDK (optional)
